Remove duplicated create-buildpacktar func
